# Приложения (code examples)/network-manager-crm-web-app/backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Mini CRM API
  description: REST API for Mini CRM Telegram Mini App
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProjectResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        clientId: { type: string, format: uuid }
        estimateId: { type: string, format: uuid }
        supervisorId: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        status: { type: string, enum: [active, done, archive] }
        type: { type: string, enum: [estimate, outstaff] }
        siteStatus: { type: string, enum: [ok, issue, downtime] }
        downtimeHours: { type: string }
        budget: { type: string }
        spent: { type: string }
        workers: { type: integer }
        confirmed: { type: integer }
        reportsOnReview: { type: integer }
    ProjectContact:
      type: object
      properties:
        role: { type: string, enum: [manager, supervisor] }
        name: { type: string }
        phone: { type: string }
    MyProjectResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }
        clientName: { type: string }
        status: { type: string, enum: [active, done, archive] }
        siteStatus: { type: string, enum: [ok, issue, downtime] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        role: { type: string, enum: [worker, supervisor] }
        confirmationStatus: { type: string, enum: [pending, confirmed, rejected] }
        assignedAt: { type: string, format: date-time }
        contacts:
          type: array
          items: { $ref: '#/components/schemas/ProjectContact' }
        projectWorkers:
          type: array
          items: { $ref: '#/components/schemas/ProjectWorkerResponse' }
    ProjectWorkerResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string, enum: [worker, supervisor] }
        confirmationStatus: { type: string, enum: [pending, confirmed, rejected] }
...
    SupervisorReportResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        projectName: { type: string }
        supervisorId: { type: string, format: uuid }
        supervisorName: { type: string }
        reportDate: { type: string, format: date }
        siteStatus: { type: string, enum: [ok, issue, downtime] }
        status: { type: string, enum: [review, accepted, attention] }
...
paths:
...
  /clients/{id}/projects:
    get:
      tags: [clients]
      security: [{ bearerAuth: [] }]
      summary: Client projects
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Project list }
...
  /projects:
    get:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: List projects
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: clientId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Project list }
    post:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Create project
      responses:
        '201': { description: Created }
  /projects/mine:
    get:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: List projects assigned to current worker/supervisor
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, done, archive] } }
        - { name: confirmationStatus, in: query, schema: { type: string, enum: [pending, confirmed] } }
      responses:
        '200':
          description: Assigned projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MyProjectResponse' }
  /projects/{id}/confirm:
    post:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Confirm project participation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [confirmed] }
  /projects/{id}/decline:
    post:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Decline project invitation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Declined
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [rejected] }
  /projects/{id}/crew:
    get:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: List project crew (supervisor only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Crew list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProjectWorkerResponse' }
        '403': { description: Forbidden for non-supervisor }
  /projects/{id}:
    get:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Get project card
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
    put:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Update project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated }
  /projects/{id}/workers:
    post:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Assign worker to project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Assigned }
  /projects/{id}/workers/{workerId}:
    delete:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Remove worker from project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: workerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Removed }
  /projects/{id}/workers/{workerId}/role:
    post:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Change worker role on project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: workerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Role updated }
  /projects/{id}/invite:
    post:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Send project invitation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Invitation sent }
...
  /reports/supervisor:
    get:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: List supervisor reports
      responses:
        '200': { description: Report list }
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Create supervisor daily report
      responses:
        '201': { description: Created }
  /reports/supervisor/{id}:
    get:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Get supervisor report
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupervisorReportResponse' }
  /reports/supervisor/{id}/transcribe:
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Transcribe voice report
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Transcription started }
  /reports/supervisor/{id}/approve:
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Approve supervisor report
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Approved }
  /reports/supervisor/{id}/attention:
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Mark supervisor report as needs attention
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated }
  /reports/supervisor/{id}/comment:
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Add manager comment
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Comment added }
...
  /finance/projects:
    get:
      tags: [finance]
      security: [{ bearerAuth: [] }]
      summary: Finance by projects
      responses:
        '200': { description: Project finance list }
  /finance/projects/{id}:
    get:
      tags: [finance]
      security: [{ bearerAuth: [] }]
      summary: Project finance details
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Details }
...